Make compatibible with OJS 3.4
